Rename "Database" namespace
