+++ Puck/GlobalUsings.cs
@@ -19,7 +19,7 @@ global using Microsoft.Extensions.Logging;
 global using Serilog;
 
 // Project namespaces.
-global using Puck.Database;
+global using Puck.Databases;
 global using Puck.Utils;
 
 // Static usings.
